[chore] 리뷰 적용
